etfdata/requirements.txt
click
finnhub-python
pandas
requests